+++ src/app/horo-common/date-time-change/date-time-change.component.ts
@@ -1,10 +1,12 @@
 import { Component, EventEmitter, OnInit, Output } from '@angular/core';
+import { addIcons } from 'ionicons';
+import { chevronBackOutline, chevronForwardOutline } from 'ionicons/icons';
 
 @Component({
-    selector: 'horo-date-time-change',
-    templateUrl: './date-time-change.component.html',
-    styleUrls: ['./date-time-change.component.scss'],
-    standalone: false
+  selector: 'horo-date-time-change',
+  templateUrl: './date-time-change.component.html',
+  styleUrls: ['./date-time-change.component.scss'],
+  standalone: false,
 })
 export class DateTimeChangeComponent implements OnInit {
   stepUnit = '年';
@@ -28,7 +30,12 @@ export class DateTimeChangeComponent implements OnInit {
     second: number;
   }>();
 
-  constructor() {}
+  constructor() {
+    addIcons({
+      chevronBackOutline,
+      chevronForwardOutline,
+    });
+  }
 
   ngOnInit() {}