+++ src/app/native/native.page.ts
@@ -4,6 +4,8 @@ import { HoroStorageService } from '../services/horostorage/horostorage.service'
 import { Horoconfig } from '../services/config/horo-config.service';
 import { Title } from '@angular/platform-browser';
 import { Path } from '../type/enum/path';
+import { deepClone } from '../utils/deep-clone';
+import { HoroRequest } from '../type/interface/request-data';
 
 @Component({
     selector: 'app-native',
@@ -13,7 +15,7 @@ import { Path } from '../type/enum/path';
 })
 export class NativePage implements OnInit {
   readonly houses: Array<string> = this.config.houses;
-  horoData = this.storage.horoData;
+  horoData: HoroRequest = deepClone(this.storage.horoData);
 
   path = Path;
   title = '本命星盘';
@@ -32,7 +34,6 @@ export class NativePage implements OnInit {
 
   getHoro() {
     this.storage.horoData = this.horoData;
-    this.storage.horoData.save();
     this.router.navigate(['./image'], { relativeTo: this.route });
   }
 }