+++ src/app/qizheng/horo/horo.component.ts
@@ -3,9 +3,13 @@ import { Title } from '@angular/platform-browser';
 import { ApiService } from 'src/app/services/api/api.service';
 import { HoroStorageService } from 'src/app/services/horostorage/horostorage.service';
 
-import * as fabric  from 'fabric';
+import * as fabric from 'fabric';
 import { Horoscope } from 'src/app/type/interface/response-qizheng';
-import { QiZhengRequst } from 'src/app/type/interface/request-data';
+import {
+  HoroRequest,
+  ProcessRequest,
+  QiZhengRequst,
+} from 'src/app/type/interface/request-data';
 import { lastValueFrom } from 'rxjs';
 import { drawHoroscope } from 'src/app/utils/image/qizheng';
 import { QizhengConfigService } from 'src/app/services/config/qizheng-config.service';
@@ -13,18 +17,21 @@ import { TipService } from 'src/app/services/qizheng/tip.service';
 import { zoomImage } from 'src/app/utils/image/horo';
 import { Platform } from '@ionic/angular';
 import { Path } from '../../type/enum/path';
+import { DeepReadonly } from 'src/app/type/interface/deep-readonly';
+import { deepClone } from 'src/app/utils/deep-clone';
 
 @Component({
-    selector: 'app-horo',
-    templateUrl: './horo.component.html',
-    styleUrls: ['./horo.component.scss'],
-    standalone: false
+  selector: 'app-horo',
+  templateUrl: './horo.component.html',
+  styleUrls: ['./horo.component.scss'],
+  standalone: false,
 })
 export class HoroComponent implements OnInit, AfterViewInit, OnDestroy {
   path = Path;
 
-  horoData = this.storage.horoData;
-  processData = this.storage.processData;
+  horoData: DeepReadonly<HoroRequest> = this.storage.horoData;
+  private processData: DeepReadonly<ProcessRequest> = this.storage.processData;
+  currentProcessData: ProcessRequest = deepClone(this.processData);
 
   horoscopeData: Horoscope | null = null;
 
@@ -50,14 +57,6 @@ export class HoroComponent implements OnInit, AfterViewInit, OnDestroy {
 
   ngOnInit() {
     this.titleService.setTitle(this.title);
-
-    // this.canvas.on('object:scaling', function(e) {
-    //   var object = e.target,
-    //       scaleX = object?.scaleX,
-    //       scaleY = object?.scaleY;
-
-    //   console.log('Object scaling. Object:', object?.get('type'), 'New scaleX:', scaleX, 'New scaleY:', scaleY);
-    // });
   }
 
   ngAfterViewInit(): void {
@@ -81,7 +80,7 @@ export class HoroComponent implements OnInit, AfterViewInit, OnDestroy {
     const requestData: QiZhengRequst = {
       native_date: this.horoData.date,
       geo: this.horoData.geo,
-      process_date: this.processData.date,
+      process_date: this.currentProcessData.date, // 使用当前的推运日期
     };
 
     try {
@@ -118,13 +117,14 @@ export class HoroComponent implements OnInit, AfterViewInit, OnDestroy {
     minute: number;
     second: number;
   }) {
+    console.log(step);
     let date = new Date(
-      this.processData.date.year,
-      this.processData.date.month - 1,
-      this.processData.date.day,
-      this.processData.date.hour,
-      this.processData.date.minute,
-      this.processData.date.second
+      this.currentProcessData.date.year,
+      this.currentProcessData.date.month - 1,
+      this.currentProcessData.date.day,
+      this.currentProcessData.date.hour,
+      this.currentProcessData.date.minute,
+      this.currentProcessData.date.second
     );
 
     date.setFullYear(date.getFullYear() + step.year);
@@ -134,12 +134,12 @@ export class HoroComponent implements OnInit, AfterViewInit, OnDestroy {
     date.setMinutes(date.getMinutes() + step.minute);
     date.setSeconds(date.getSeconds() + step.second);
 
-    this.processData.date.year = date.getFullYear();
-    this.processData.date.month = date.getMonth() + 1;
-    this.processData.date.day = date.getDate();
-    this.processData.date.hour = date.getHours();
-    this.processData.date.minute = date.getMinutes();
-    this.processData.date.second = date.getSeconds();
+    this.currentProcessData.date.year = date.getFullYear();
+    this.currentProcessData.date.month = date.getMonth() + 1;
+    this.currentProcessData.date.day = date.getDate();
+    this.currentProcessData.date.hour = date.getHours();
+    this.currentProcessData.date.minute = date.getMinutes();
+    this.currentProcessData.date.second = date.getSeconds();
 
     await this.drawHoroscope();
   }