+++ CHANGELOG.md
@@ -1,8 +1,22 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+
+### Planned
+
+- 增加十干化曜功能
+- 推运计划增加次限、三限
+
 ### Changed
+
 - Upgraded Angular to v20.0.5
 - Upgraded fabric to v6.7.0
 - Upgraded ionicons to v8.0.9
+- archive 组件 的 edit 子组件中姓名字段长度调整到 30 个字符
+
+### Fixed
+
+- 重构 horostorage.service，以修复 image 组件保存 Note 时的错误
+- 修复 qizheng 天宫图中多颗行星相距太近指示线错乱